Crate socketioxide_core

This crate is the core of the socketioxide crate. It contains basic types and interfaces for the socketioxide crate and the parser sub-crates.

Modules

packet

Socket.io packet implementation. The Packet is the base unit of data that is sent over the engine.io socket.

parser

Contains all the type and interfaces for the parser sub-crates.

Structs

Sid

A 128 bit session id type representing a base64 16 char string

Str

A custom Bytes wrapper to efficiently store string packets

Enums

Value

Represents a value that can be sent over the engine.io wire as an engine.io packet or the data that can be outputed by a binary parser (e.g. MsgPackParser) or a string parser (e.g. CommonParser)).